simplify structure
