--- blocking/src/main/kotlin/com/example/http/capacity/simulator/blocking/app/BlockingApplication.kt
@@ -1,6 +1,0 @@
-package com.example.http.capacity.simulator.blocking.app
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication(scanBasePackages = ["com.example.http.capacity.simulator.blocking"])
-class BlockingApplication
