--- blocking/src/main/kotlin/com/example/http/capacity/simulator/blocking/config/Config.kt
@@ -1,17 +1,0 @@
-package com.example.http.capacity.simulator.blocking.config
-
-import org.springframework.boot.web.client.RestTemplateBuilder
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.client.RestTemplate
-
-@Configuration
-class Config {
-    @Bean
-    fun restTemplate(builder: RestTemplateBuilder): RestTemplate =
-        builder.rootUri(baseUrl).build()
-
-    companion object {
-        const val baseUrl = "https://postman-echo.com/delay"
-    }
-}
