+++ build.gradle.kts
@@ -1,34 +1,41 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-import org.springframework.boot.gradle.tasks.bundling.BootJar
 
 plugins {
     id("org.springframework.boot") version "2.3.4.RELEASE"
     id("io.spring.dependency-management") version "1.0.10.RELEASE"
     kotlin("jvm") version "1.4.10"
     kotlin("plugin.spring") version "1.4.10"
+    id("com.bmuschko.docker-spring-boot-application") version "6.6.1"
 }
 
 group = "com.example"
 version = "0.0.1-SNAPSHOT"
 java.sourceCompatibility = JavaVersion.VERSION_11
 
-allprojects {
-
-    apply(plugin = "kotlin")
-    apply(plugin = "kotlin-spring")
-
-    tasks.withType<BootJar> {
-        mainClassName = "com.example.http.capacity.simulator.launcher.LauncherKt.main"
+tasks.withType<KotlinCompile> {
+    kotlinOptions {
+        freeCompilerArgs = listOf("-Xjsr305=strict")
+        jvmTarget = "11"
     }
+}
 
-    tasks.withType<KotlinCompile> {
-        kotlinOptions {
-            freeCompilerArgs = listOf("-Xjsr305=strict")
-            jvmTarget = "11"
-        }
-    }
+repositories {
+    mavenCentral()
+}
 
-    repositories {
-        mavenCentral()
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-webflux")
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.11.3")
+
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor")
+}
+
+docker {
+    springBootApplication {
+        baseImage.set("adoptopenjdk/openjdk11")
+        ports.set(listOf(8080))
+        images.set(setOf("${rootProject.name}:latest"))
+        jvmArgs.set(listOf("-Xmx2048m"))
     }
 }
