--- launcher/build.gradle.kts
@@ -1,7 +1,0 @@
-dependencies {
-    implementation(project(":blocking"))
-    implementation(project(":nonblocking"))
-
-    implementation("org.springframework.boot:spring-boot-starter")
-    implementation("commons-cli:commons-cli:1.4")
-}
