--- nonblocking/src/main/kotlin/com/example/http/capacity/simulator/nonblocking/app/NonBlockingApplication.kt
@@ -1,6 +1,0 @@
-package com.example.http.capacity.simulator.nonblocking.app
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication(scanBasePackages = ["com.example.http.capacity.simulator.nonblocking"])
-class NonBlockingApplication
