--- nonblocking/src/main/resources/non-blocking.http
@@ -1,4 +1,0 @@
-GET http://localhost:8090/api/nonblocking
-Accept: application/json
-
-###
