+++ settings.gradle.kts
@@ -1,7 +1,3 @@
 apply(from = "./resolutions.gradle.kts")
 
 rootProject.name = "http-capacity-simulator"
-
-include(":launcher")
-include(":blocking")
-include(":nonblocking")
