+++ src/main/kotlin/com/example/http/capacity/simulator/launcher/Launcher.kt
@@ -1,0 +1,11 @@
+package com.example.http.capacity.simulator.launcher
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication(scanBasePackages = ["com.example.http.capacity.simulator"])
+class Launcher
+
+fun main() {
+    runApplication<Launcher>()
+}
